fixed logout

+++ front-end/my-app/src/components/nav/nav.js
@@ -1,11 +1,22 @@
-import React from 'react';
+import React, {useState} from 'react';
 import './nav.css';
 import { useHistory, Link } from "react-router-dom";
+import service from '../../service/service';
+
+
+import {useDispatch, useSelector} from 'react-redux';
+import { Sigin } from '../../redux/action/siginAction';
+     
 const NavBar = () => {
     let refreshToken =  localStorage.getItem("refreshToken") || "";
-
-    const handleLogout = () => {
-        localStorage.clear();
+    const [route, setRoute] = useState(false);
+    const handleLogout = async () => {
+        console.log(refreshToken);
+        let response = await service.fetchLogoutRoute(refreshToken);
+        if(response?.status === 200){
+            localStorage.clear();
+            setRoute(true)
+        }
     }
     return (
         <nav id="nav-bar">
